Week16/operational_agent_api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.9
langchain-openai==0.0.8
langchain-core==0.1.26
langsmith==0.0.87
pydantic==2.6.2
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.2
